Fix instance shutdown never timing out
In an attempt to try and prevent terminating an instance while a CI task
is running, the shutdown script checks for the existance of an agent
process.  Previously a calculation of a timeout for this delay was
stored, however it was never actually used.  Fix this by aborting the
delay after the timeout has expired.

Signed-off-by: Chris Evich <cevich@redhat.com>